Refresh TDH replacement snapshot before persistence
