+++ pr-reports/1862.md
@@ -13,14 +13,16 @@ current confirmed clusters.
 
 ## What Changed
 
-- Reuses the current consolidated TDH snapshot to recover old memberships and
-  determine the exact derived rows being replaced.
+- Uses the pre-update consolidated TDH snapshot to recover old memberships,
+  then refreshes it immediately before rebuilding to determine the exact
+  derived rows being replaced.
 - Resolves current confirmed clusters through bounded, parameterized batches.
 - Deletes the exact affected consolidation keys from consolidated TDH, historic,
   season/edition, and per-NFT TDH data before rebuilding the complete
   replacement set.
 - Rejects a proposed final state that assigns one wallet to multiple
-  consolidation rows.
+  consolidation rows, including rows whose wallet list is JSON encoded at
+  runtime.
 - Covers the A-B to A-C/B and back to A-B regression, including event
   direction, casing, multi-event batches, combined TDH, and stale standalone
   membership.
